Refactor Length class to module and rename to convert
Avoids confusion of initialising single instance and passing around. All important data for calculating conversions and aliases was already stored as global constants in the module, so now just use these directly from functions instead of incorporating into class
